requirements.txt
flask==2.3.3
websockets==11.0.3
google-cloud-speech==2.21.0
google-cloud-texttospeech==2.14.1
google-cloud-translate==3.11.1
pyaudio==0.2.11
requests==2.31.0
